requirements.txt
tensorflow
numpy
Pillow
scikit-learn